src/ml/requirements.txt
# Core dependencies
pecanpy>=2.0.0
gensim>=4.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
scipy>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-ai>=0.0.14

# Optional (for advanced features)
umap-learn>=0.5.4
plotly>=5.17.0
